Add create persona session API to create a snallabot session (#3)
Added a createPersonaSession API. This returns an authenticated
sessionId that clients can use to connect to EA servers. Every client
wanting to use snallabot to connect to EA should do this once per user
and persist the session id.

For database, I am using firebase again. This firebase DB will be shared
among all services, but it is expected that each service queries
different collections. This is to keep concerns separate. For
development, I already setup the firebase emulator so there is no need
to use any production DB!